setup.v2.sh: docker-compose 파일 위치 변경 (#75)
## 변경사항
- dl.querypie.com 에서 설치하는 docker-compose 의 파일위치를 변경합니다.
- 기존: `https://dl.querypie.com/releases/bin/docker-compose-$(uname
-s)-$(uname -m)`
- 변경:
`https://dl.querypie.com/releases/bin/v2.39.1/docker-compose-${kernel}-${hardware}`
- macos 에서 인식하는 hardware platform 이 arm64 인 경우, aarch64 로 값을 바꿉니다.
- install::docker_compose 를 리팩토링합니다.
  - 설치된, 또는 설치한 docker-compose 버전을 보여줍니다.
  - 로깅 메시지를 개선합니다.

## 테스팅 결과
- macbook 에서 설치에 성공합니다. `./setup.v2.sh --install 11.0.1-beta`
- macbook 에서 Docker Desktop 패키지 중 일부인 /usr/local/bin/docker-compose 를
삭제하고, 설치에 성공합니다.
https://dl.querypie.com/releases/bin/v2.39.1/docker-compose-darwin-aarch64
를 설치하고, 정상 작동합니다.
- Packer 스크립트, `./ubuntu24.04-install.sh 11.0.1` 를 성공합니다.
- Packer 스크립트, `./az2023-install.sh 11.0.1` 를 성공합니다.

+++ aws-ami/scripts/setup.v2.sh
@@ -241,29 +241,39 @@ function install::docker_compose() {
   if [[ ${COMPOSE} == docker-compose ]]; then
     if command_exists docker-compose; then
       echo >&2 "# Docker Compose is already installed at $(command -v docker-compose)"
+      log::do docker-compose --version
       return
     else
-      echo >&2 "# Docker Compose is not installed. Installing now."
-      log::do curl -fsSL "https://dl.querypie.com/releases/bin/docker-compose-$(uname -s)-$(uname -m)" -o docker-compose
-      log::sudo install -m 755 docker-compose /usr/local/bin
-      rm docker-compose
-      return
+      local kernel hardware
+      kernel=$(uname -s | tr '[:upper:]' '[:lower:]')
+      hardware=$(uname -m | tr '[:upper:]' '[:lower:]')
+      [[ $hardware == arm64 ]] && hardware=aarch64
+      if [[ $kernel == linux || $kernel == darwin ]] && [[ $hardware == x86_64 || $hardware == aarch64 ]]; then
+        echo >&2 "# Docker Compose is not installed. Installing now."
+        log::do curl -fsSL "https://dl.querypie.com/releases/bin/v2.39.1/docker-compose-${kernel}-${hardware}" -o docker-compose
+        log::sudo install -m 755 docker-compose /usr/local/bin
+        rm docker-compose
+        log::do docker-compose --version
+        return
+      fi
     fi
   elif [[ ${COMPOSE} == podman-compose ]]; then
     if command_exists podman-compose; then
       echo >&2 "# Podman Compose is already installed at $(command -v podman-compose)"
+      log::do podman-compose --version
       return
     else
       echo >&2 "# Podman Compose is not installed. Please refer to the installation manual."
       log::error "Please report this problem to the technical support team of QueryPie."
       exit 1
     fi
-  else
-    echo >&2 "# Docker Compose is not installed. Unknown version of Docker."
-    log::do docker --version
-    log::error "Please report this problem to the technical support team of QueryPie."
-    exit 1
   fi
+
+  echo >&2 "# Docker Compose is not installed. It seems to be an unsupported platform."
+  log::do uname -a
+  log::do ${DOCKER} --version
+  log::error "Please report this problem to the technical support team of QueryPie."
+  exit 1
 }
 
 function install::config_files() {
